Make date parsing much stricter and move date util functions into their own module
