+++ temba/utils/tests.py
@@ -29,12 +29,12 @@ from temba.orgs.models import Org, UserSettings
 from temba.tests import TembaTest
 from temba_expressions.evaluator import EvaluationContext, DateStyle
 
-from . import format_decimal, str_to_datetime, str_to_time, date_to_utc_range
-from . import json_to_dict, dict_to_struct, datetime_to_ms, ms_to_datetime, dict_to_json, str_to_bool
-from . import percentage, datetime_to_json_date, json_date_to_datetime
-from . import datetime_to_str, chunk_list, get_country_code_by_name, datetime_to_epoch, voicexml
+from . import format_decimal, json_to_dict, dict_to_struct, dict_to_json, str_to_bool, percentage, datetime_to_json_date
+from . import chunk_list, get_country_code_by_name, voicexml, json_date_to_datetime
 from .cache import get_cacheable_result, get_cacheable_attr, incrby_existing, QueueRecord
 from .currencies import currency_for_country
+from .dates import str_to_datetime, str_to_time, date_to_utc_range, datetime_to_ms, ms_to_datetime, datetime_to_epoch
+from .dates import datetime_to_str
 from .email import send_simple_email, is_valid_address
 from .export import TableExporter
 from .expressions import migrate_template, evaluate_template, evaluate_template_compat, get_function_listing
@@ -84,110 +84,6 @@ class InitTest(TembaTest):
 
             self.assertEqual('Tm93IGlzDQp0aGUgdGltZQ0KZm9yIGFsbCBnb29kDQpwZW9wbGUgdG8NCnJlc2lzdC4NCg0KSG93IGFib3V0IGhhaWt1cz8NCkkgZmluZCB0aGVtIHRvIGJlIGZyaWVuZGx5Lg0KcmVmcmlnZXJhdG9yDQoNCjAxMjM0NTY3ODkNCiFAIyQlXiYqKCkgW117fS09Xys7JzoiLC4vPD4/fFx+YA0KQUJDREVGR0hJSktMTU5PUFFSU1RVVldYWVphYmNkZWZnaGlqa2xtbm9wcXJzdHV2d3h5eg==',
                              decode_base64('Tm93IGlzDQp0aGUgdGltZQ0KZm9yIGFsbCBnb29kDQpwZW9wbGUgdG8NCnJlc2lzdC4NCg0KSG93IGFib3V0IGhhaWt1cz8NCkkgZmluZCB0aGVtIHRvIGJlIGZyaWVuZGx5Lg0KcmVmcmlnZXJhdG9yDQoNCjAxMjM0NTY3ODkNCiFAIyQlXiYqKCkgW117fS09Xys7JzoiLC4vPD4/fFx+YA0KQUJDREVGR0hJSktMTU5PUFFSU1RVVldYWVphYmNkZWZnaGlqa2xtbm9wcXJzdHV2d3h5eg=='))
-
-    def test_datetime_to_ms(self):
-        d1 = datetime.datetime(2014, 1, 2, 3, 4, 5, tzinfo=pytz.utc)
-        self.assertEqual(datetime_to_ms(d1), 1388631845000)  # from http://unixtimestamp.50x.eu
-        self.assertEqual(ms_to_datetime(1388631845000), d1)
-
-        tz = pytz.timezone("Africa/Kigali")
-        d2 = tz.localize(datetime.datetime(2014, 1, 2, 3, 4, 5))
-        self.assertEqual(datetime_to_ms(d2), 1388624645000)
-        self.assertEqual(ms_to_datetime(1388624645000), d2.astimezone(pytz.utc))
-
-    def test_datetime_to_json_date(self):
-        d1 = datetime.datetime(2014, 1, 2, 3, 4, 5, tzinfo=pytz.utc)
-        self.assertEqual(datetime_to_json_date(d1), '2014-01-02T03:04:05.000Z')
-        self.assertEqual(json_date_to_datetime('2014-01-02T03:04:05.000Z'), d1)
-        self.assertEqual(json_date_to_datetime('2014-01-02T03:04:05.000'), d1)
-
-        tz = pytz.timezone("Africa/Kigali")
-        d2 = tz.localize(datetime.datetime(2014, 1, 2, 3, 4, 5))
-        self.assertEqual(datetime_to_json_date(d2), '2014-01-02T01:04:05.000Z')
-        self.assertEqual(json_date_to_datetime('2014-01-02T01:04:05.000Z'), d2.astimezone(pytz.utc))
-        self.assertEqual(json_date_to_datetime('2014-01-02T01:04:05.000'), d2.astimezone(pytz.utc))
-
-    def test_datetime_to_str(self):
-        tz = pytz.timezone("Africa/Kigali")
-        d2 = tz.localize(datetime.datetime(2014, 1, 2, 3, 4, 5, 6))
-
-        self.assertEqual(datetime_to_str(d2), '2014-01-02T01:04:05.000006Z')  # no format
-        self.assertEqual(datetime_to_str(d2, format='%Y-%m-%d'), '2014-01-02')  # format provided
-        self.assertEqual(datetime_to_str(d2, tz=tz), '2014-01-02T03:04:05.000006Z')  # in specific timezone
-        self.assertEqual(datetime_to_str(d2, ms=False), '2014-01-02T01:04:05Z')  # no ms
-        self.assertEqual(datetime_to_str(d2.date()), '2014-01-02T00:00:00.000000Z')  # no ms
-
-    def test_datetime_to_epoch(self):
-        dt = json_date_to_datetime('2014-01-02T01:04:05.000Z')
-        self.assertEqual(1388624645, datetime_to_epoch(dt))
-
-    def test_str_to_datetime(self):
-        tz = pytz.timezone('Asia/Kabul')
-        with patch.object(timezone, 'now', return_value=tz.localize(datetime.datetime(2014, 1, 2, 3, 4, 5, 6))):
-            self.assertIsNone(str_to_datetime(None, tz))  # none
-            self.assertIsNone(str_to_datetime('', tz))  # empty string
-            self.assertIsNone(str_to_datetime('xxx', tz))  # unparseable string
-            self.assertIsNone(str_to_datetime('xxx', tz, fill_time=False))  # unparseable string
-            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 3, 4, 5, 6)),
-                             str_to_datetime('01-02-2013', tz, dayfirst=True))  # day first
-            self.assertEqual(tz.localize(datetime.datetime(2013, 1, 2, 3, 4, 5, 6)),
-                             str_to_datetime('01-02-2013', tz, dayfirst=False))  # month first
-            self.assertEqual(tz.localize(datetime.datetime(2013, 1, 31, 3, 4, 5, 6)),
-                             str_to_datetime('01-31-2013', tz, dayfirst=True))  # impossible as day first
-            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 5, 6)),
-                             str_to_datetime('01-02-2013 07:08', tz, dayfirst=True))  # hour and minute provided
-            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000)),
-                             str_to_datetime('01-02-2013 07:08:09.100000', tz, dayfirst=True))  # complete time provided
-            self.assertEqual(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000, tzinfo=pytz.UTC),
-                             str_to_datetime('01-02-2013 07:08:09.100000Z', tz, dayfirst=True))  # Z marker
-            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000)),
-                             str_to_datetime('2013-02-01T07:08:09.100000+04:30', tz, dayfirst=True))  # ISO in local tz
-            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000)),
-                             str_to_datetime('2013-02-01T04:38:09.100000+02:00', tz, dayfirst=True))  # ISO in other tz
-            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000)),
-                             str_to_datetime('2013-02-01T00:38:09.100000-02:00', tz, dayfirst=True))  # ISO in other tz
-            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000)),
-                             str_to_datetime('2013-02-01T07:08:09.100000+04:30.', tz, dayfirst=True))  # trailing period
-            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 0, 0, 0, 0)),
-                             str_to_datetime('01-02-2013', tz, dayfirst=True, fill_time=False))  # no time filling
-
-            # just year
-            self.assertEqual(datetime.datetime(123, 1, 2, 3, 4, 5, 6, tz),
-                             str_to_datetime('123', tz))
-
-        # localizing while in DST to something outside DST
-        tz = pytz.timezone('US/Eastern')
-        with patch.object(timezone, 'now', return_value=tz.localize(datetime.datetime(2029, 11, 1, 12, 30, 0, 0))):
-            parsed = str_to_datetime('06-11-2029', tz, dayfirst=True)
-            self.assertEqual(tz.localize(datetime.datetime(2029, 11, 6, 12, 30, 0, 0)),
-                             parsed)
-
-            # assert there is no DST offset
-            self.assertFalse(parsed.tzinfo.dst(parsed))
-
-            self.assertEqual(tz.localize(datetime.datetime(2029, 11, 6, 13, 45, 0, 0)),
-                             str_to_datetime('06-11-2029 13:45', tz, dayfirst=True))
-
-        # deal with datetimes that have timezone info
-        self.assertEqual(pytz.utc.localize(datetime.datetime(2016, 11, 21, 20, 36, 51, 215681)).astimezone(tz),
-                         str_to_datetime('2016-11-21T20:36:51.215681Z', tz))
-
-        self.assertEqual(datetime.datetime(123, 1, 2, 5, 4, 5, 6, pytz.utc),
-                         str_to_datetime('123-1-2T5:4:5.000006Z', tz))
-
-    def test_str_to_time(self):
-        tz = pytz.timezone('Asia/Kabul')
-        with patch.object(timezone, 'now', return_value=tz.localize(datetime.datetime(2014, 1, 2, 3, 4, 5, 6))):
-            self.assertEqual(datetime.time(3, 4), str_to_time('03:04'))  # zero padded
-            self.assertEqual(datetime.time(3, 4), str_to_time('3:4'))  # not zero padded
-            self.assertEqual(datetime.time(3, 4), str_to_time('01-02-2013 03:04'))  # with date
-            self.assertEqual(datetime.time(15, 4), str_to_time('3:04 PM'))  # as PM
-
-    def test_date_to_utc_range(self):
-        self.assertEqual(date_to_utc_range(datetime.date(2017, 2, 20), self.org), (
-            datetime.datetime(2017, 2, 19, 22, 0, 0, 0, tzinfo=pytz.UTC),
-            datetime.datetime(2017, 2, 20, 22, 0, 0, 0, tzinfo=pytz.UTC)
-        ))
 
     def test_str_to_bool(self):
         self.assertFalse(str_to_bool(None))
@@ -252,6 +148,103 @@ class InitTest(TembaTest):
 
         self.assertEqual(headers, {'User-agent': 'RapidPro', 'Foo': "Bar", 'Token': "123456"})
         self.assertEqual(http_headers(), {'User-agent': 'RapidPro'})  # check changes don't leak
+
+
+class DatesTest(TembaTest):
+    def test_datetime_to_ms(self):
+        d1 = datetime.datetime(2014, 1, 2, 3, 4, 5, tzinfo=pytz.utc)
+        self.assertEqual(datetime_to_ms(d1), 1388631845000)  # from http://unixtimestamp.50x.eu
+        self.assertEqual(ms_to_datetime(1388631845000), d1)
+
+        tz = pytz.timezone("Africa/Kigali")
+        d2 = tz.localize(datetime.datetime(2014, 1, 2, 3, 4, 5))
+        self.assertEqual(datetime_to_ms(d2), 1388624645000)
+        self.assertEqual(ms_to_datetime(1388624645000), d2.astimezone(pytz.utc))
+
+    def test_datetime_to_json_date(self):
+        d1 = datetime.datetime(2014, 1, 2, 3, 4, 5, tzinfo=pytz.utc)
+        self.assertEqual(datetime_to_json_date(d1), '2014-01-02T03:04:05.000Z')
+        self.assertEqual(json_date_to_datetime('2014-01-02T03:04:05.000Z'), d1)
+        self.assertEqual(json_date_to_datetime('2014-01-02T03:04:05.000'), d1)
+
+        tz = pytz.timezone("Africa/Kigali")
+        d2 = tz.localize(datetime.datetime(2014, 1, 2, 3, 4, 5))
+        self.assertEqual(datetime_to_json_date(d2), '2014-01-02T01:04:05.000Z')
+        self.assertEqual(json_date_to_datetime('2014-01-02T01:04:05.000Z'), d2.astimezone(pytz.utc))
+        self.assertEqual(json_date_to_datetime('2014-01-02T01:04:05.000'), d2.astimezone(pytz.utc))
+
+    def test_datetime_to_str(self):
+        tz = pytz.timezone("Africa/Kigali")
+        d2 = tz.localize(datetime.datetime(2014, 1, 2, 3, 4, 5, 6))
+
+        self.assertEqual(datetime_to_str(d2), '2014-01-02T01:04:05.000006Z')  # no format
+        self.assertEqual(datetime_to_str(d2, format='%Y-%m-%d'), '2014-01-02')  # format provided
+        self.assertEqual(datetime_to_str(d2, tz=tz), '2014-01-02T03:04:05.000006Z')  # in specific timezone
+        self.assertEqual(datetime_to_str(d2, ms=False), '2014-01-02T01:04:05Z')  # no ms
+        self.assertEqual(datetime_to_str(d2.date()), '2014-01-02T00:00:00.000000Z')  # no ms
+
+    def test_datetime_to_epoch(self):
+        dt = json_date_to_datetime('2014-01-02T01:04:05.000Z')
+        self.assertEqual(1388624645, datetime_to_epoch(dt))
+
+    def test_str_to_datetime(self):
+        tz = pytz.timezone('Asia/Kabul')
+        with patch.object(timezone, 'now', return_value=tz.localize(datetime.datetime(2014, 1, 2, 3, 4, 5, 6))):
+            self.assertIsNone(str_to_datetime(None, tz))  # none
+            self.assertIsNone(str_to_datetime('', tz))  # empty string
+            self.assertIsNone(str_to_datetime('xxx', tz))  # unparseable string
+            self.assertIsNone(str_to_datetime('xxx', tz, fill_time=False))  # unparseable string
+            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 3, 4, 5, 6)),
+                             str_to_datetime('01-02-2013', tz, dayfirst=True))  # day first
+            self.assertEqual(tz.localize(datetime.datetime(2013, 1, 2, 3, 4, 5, 6)),
+                             str_to_datetime('01-02-2013', tz, dayfirst=False))  # month first
+            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 0, 0)),
+                             str_to_datetime('01-02-2013 07:08', tz, dayfirst=True))  # hour and minute provided
+            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000)),
+                             str_to_datetime('01-02-2013 07:08:09.100000', tz, dayfirst=True))  # complete time provided
+            self.assertEqual(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000, tzinfo=pytz.UTC),
+                             str_to_datetime('2013-02-01T07:08:09.100000Z', tz, dayfirst=True))  # Z marker
+            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000)),
+                             str_to_datetime('2013-02-01T07:08:09.100000+04:30', tz, dayfirst=True))  # ISO in local tz
+            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000)),
+                             str_to_datetime('2013-02-01T04:38:09.100000+02:00', tz, dayfirst=True))  # ISO in other tz
+            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000)),
+                             str_to_datetime('2013-02-01T00:38:09.100000-02:00', tz, dayfirst=True))  # ISO in other tz
+            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 7, 8, 9, 100000)),
+                             str_to_datetime('2013-02-01T07:08:09.100000+04:30.', tz, dayfirst=True))  # trailing period
+            self.assertEqual(tz.localize(datetime.datetime(2013, 2, 1, 0, 0, 0, 0)),
+                             str_to_datetime('01-02-2013', tz, dayfirst=True, fill_time=False))  # no time filling
+
+        # localizing while in DST to something outside DST
+        tz = pytz.timezone('US/Eastern')
+        with patch.object(timezone, 'now', return_value=tz.localize(datetime.datetime(2029, 11, 1, 12, 30, 0, 0))):
+            parsed = str_to_datetime('06-11-2029', tz, dayfirst=True)
+            self.assertEqual(tz.localize(datetime.datetime(2029, 11, 6, 12, 30, 0, 0)),
+                             parsed)
+
+            # assert there is no DST offset
+            self.assertFalse(parsed.tzinfo.dst(parsed))
+
+            self.assertEqual(tz.localize(datetime.datetime(2029, 11, 6, 13, 45, 0, 0)),
+                             str_to_datetime('06-11-2029 13:45', tz, dayfirst=True))
+
+        # deal with datetimes that have timezone info
+        self.assertEqual(pytz.utc.localize(datetime.datetime(2016, 11, 21, 20, 36, 51, 215681)).astimezone(tz),
+                         str_to_datetime('2016-11-21T20:36:51.215681Z', tz))
+
+    def test_str_to_time(self):
+        tz = pytz.timezone('Asia/Kabul')
+        with patch.object(timezone, 'now', return_value=tz.localize(datetime.datetime(2014, 1, 2, 3, 4, 5, 6))):
+            self.assertEqual(datetime.time(3, 4), str_to_time('03:04'))  # zero padded
+            self.assertEqual(datetime.time(3, 4), str_to_time('3:4'))  # not zero padded
+            self.assertEqual(datetime.time(3, 4), str_to_time('01-02-2013 03:04'))  # with date
+            self.assertEqual(datetime.time(15, 4), str_to_time('3:04 PM'))  # as PM
+
+    def test_date_to_utc_range(self):
+        self.assertEqual(date_to_utc_range(datetime.date(2017, 2, 20), self.org), (
+            datetime.datetime(2017, 2, 19, 22, 0, 0, 0, tzinfo=pytz.UTC),
+            datetime.datetime(2017, 2, 20, 22, 0, 0, 0, tzinfo=pytz.UTC)
+        ))
 
 
 class TimezonesTest(TembaTest):
